Add std versions of logs & telemetry for the host side
The versions with heapless linear maps and especially static strings
are difficult to work with, but because Postcard works with raw serde
types internally they are equivalent to std HashMaps and Strings.

This commit duplicates the existing Log and TelemetryFrame data
structures with std drop-ins, which can be easily deserialized and
worked with afterwards. As a bonus, they support arbitrary sizes.

It's not the most elegant solution, but it works for the time being.